services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
ultralytics==8.3.55
pillow==11.1.0
numpy==2.2.1
arabic-reshaper==3.0.0
python-bidi==0.6.3
httpx==0.28.1
pytest==8.3.4
